pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Django REST API backend"
requires-python = ">=3.11"

dependencies = [
    "django>=5.2.10",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "pytz>=2025.2",
]
